stop

+++ amigamame106/amiga106_config.h
@@ -226,7 +226,7 @@ public:
         std::string _romsPath,_userPath;
         bool    _useCheatCodeFile = false;
         std::string _cheatFilePath="PROGDIR:cheat.dat";
-        int     _speedlimit = 100;
+        float     _speedlimit = 100.0f;
 
     };
     Misc &misc() { return _misc; }
